# backend/internal/api/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Container-scoped firewall control plane. Authentication is bearer-token on TCP listeners and peer-credentials on unix sockets.",
        "title": "Firefik API",
        "contact": {},
        "license": {
            "name": "Apache-2.0"
        },
        "version": "0.3"
    },
    "basePath": "/",
    "paths": {
        "/api/agents": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies GET /v1/agents on the configured control-plane.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet"
                ],
                "summary": "List fleet agents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/agents/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies GET /v1/agents/{id} on the configured control-plane.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet"
                ],
                "summary": "Get agent detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "agent instance id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/agents/{id}/commands": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies POST /v1/agents/{id}/commands on the configured control-plane.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet"
                ],
                "summary": "Send command to an agent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "agent instance id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "command payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/agents/{id}/snapshot": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies GET /v1/agents/{id}/snapshot on the configured control-plane.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fleet"
                ],
                "summary": "Get agent latest snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "agent instance id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/approvals": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies GET /v1/approvals on the configured control-plane.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "approvals"
                ],
                "summary": "List approvals",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies POST /v1/approvals on the configured control-plane.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "approvals"
                ],
                "summary": "Create an approval request",
                "parameters": [
                    {
                        "description": "approval payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/approvals/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies GET /v1/approvals/{id} on the configured control-plane.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "approvals"
                ],
                "summary": "Get an approval",
                "parameters": [
                    {
                        "type": "string",
                        "description": "approval ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/approvals/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies POST /v1/approvals/{id}/approve on the configured control-plane.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "approvals"
                ],
                "summary": "Approve an approval request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "approval ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "approval decision payload",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/approvals/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Proxies POST /v1/approvals/{id}/reject on the configured control-plane.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "approvals"
                ],
                "summary": "Reject an approval request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "approval ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "rejection payload",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/audit/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the most recent audit events kept in the in-memory ring buffer. Populated only when `history` is included in `FIREFIK_AUDIT_SINK`. Supports `?limit=\u003cn\u003e` and `?since=\u003cRFC3339\u003e`.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Audit history snapshot",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "maximum events to return (0 = all)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 timestamp; only events strictly after this are returned",
                        "name": "since",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/autogen/proposals": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns observe-mode proposals per container. Populated only when `FIREFIK_AUTOGEN_MODE=observe`.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "autogen"
                ],
                "summary": "Auto-generated rule proposals",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
        "/api/autogen/proposals/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Materialises a pending observe-mode proposal as labels or a policy snippet and records the approval.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "autogen"
                ],
                "summary": "Approve an autogen proposal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "container ID or unambiguous prefix",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "approval mode (`labels` or `policy`)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.approveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.approveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/autogen/proposals/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Marks an observe-mode proposal as rejected with an optional reason.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "autogen"
                ],
                "summary": "Reject an autogen proposal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "container ID or unambiguous prefix",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "rejection reason",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.rejectRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
        "/api/containers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns every container visible to the Docker daemon with its parsed firewall configuration and live apply status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "containers"
                ],
                "summary": "List containers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.ContainerDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_api.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.approveRequest": {
            "type": "object",
            "properties": {
                "mode": {
                    "type": "string"
                }
            }
        },
        "internal_api.approveResponse": {
            "type": "object",
            "properties": {
                "container_id": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "peers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ports": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "snippet": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}